fix[FoundationPoseRender]: fix parameter type error in `ceil_div`

+++ detection_6d_foundationpose/src/foundationpose_render.cu
@@ -340,7 +340,7 @@ __global__ void transform_points_kernel(const float *transform_matrixs,
   transformed_point[2] = matrix[2] * x + matrix[6] * y + matrix[10] * z + matrix[14];
 }
 
-static uint16_t ceil_div(uint16_t numerator, uint16_t denominator)
+static uint32_t ceil_div(uint32_t numerator, uint32_t denominator)
 {
   uint32_t accumulator = numerator + denominator - 1;
   return accumulator / denominator + 1;
